# Conflicts:
#	src/main/java/ee/coffee/coffeeshop/controllers/OrderController.java
#	src/main/java/ee/coffee/coffeeshop/models/enums/OrderStatus.java
#	src/main/resources/templates/products.ftlh

+++ src/main/java/ee/coffee/coffeeshop/services/interfaces/CartService.java
@@ -1,12 +1,7 @@
 package ee.coffee.coffeeshop.services.interfaces;
 
-import ee.coffee.coffeeshop.entity.Cart;
-import jakarta.transaction.Transactional;
-
-import java.util.List;
-
 public interface CartService {
-    void addProductToCart(Integer productId, Integer quantity, Integer userId);
+    void addProductToCart(Long productId, Integer quantity, Integer userId);
 
     void deleteProductFromCart(Integer userId, Integer productId);
